+++ src/commands/users/fishy.rs
@@ -31,7 +31,7 @@ async fn fishy(ctx: &Context, msg: &Message, args: Args) -> CommandResult {
 
     let is_self = target_id == msg.author.id;
 
-    let author_user_data = UserData::from_id_or_new(&ctx, msg.author.id).await?;
+    let mut author_user_data = UserData::from_id_or_new(&ctx, msg.author.id).await?;
 
     if let Some(duration_str) = author_user_data.fishies_humantime_cooldown() {
         msg.channel_id
@@ -47,21 +47,48 @@ async fn fishy(ctx: &Context, msg: &Message, args: Args) -> CommandResult {
         None
     };
 
-    match target_user_data {
-        Some(target) => {
+    let (fishies, is_golden) = match target_user_data {
+        Some(mut target) => {
             // Someone else
-            target.inc_fishies(is_self).save(&ctx).await?;
+            let fishies_tup = target.inc_fishies(is_self);
+            target.save(&ctx).await?;
             // So we need to save author separately
             author_user_data.reset_last_rep().save(&ctx).await?;
+
+            fishies_tup
         }
         None => {
-            author_user_data
-                .inc_fishies(is_self)
-                .reset_last_fishy()
+            let fishies_tup = author_user_data.inc_fishies(is_self);
+
+            author_user_data.reset_last_fishy()
                 .save(&ctx)
                 .await?;
+            
+            fishies_tup
         }
-    }
+    };
+
+    let target_user_tag = if is_self {
+        msg.author.tag()
+    } else {
+        target_id.to_user(&ctx).await?.face()
+    };
+
+    let name_str = if is_self {
+        "".to_string()
+    } else {
+        format!(" for {}", target_user_tag)
+    };
+
+    let s = if is_golden {
+        format!("You caught a golden fishy{}!!! ({} fishies)", name_str, fishies)
+    } else {
+        format!("You caught {} fishies{}!", fishies, name_str)
+    };
+
+    msg.channel_id
+        .say(&ctx, s)
+        .await?;
 
     Ok(())
 }